requirements.txt
# One NZ Employee Finder Agent - Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Database
# SQLite is built into Python, no additional package needed

# Data Processing
pandas==2.1.4
openpyxl==3.1.2  # For Excel file reading

# Utilities
python-multipart==0.0.6  # For file uploads
python-dotenv==1.0.0  # For environment variables
requests==2.31.0  # For LLM API calls

# Optional: For enhanced NLP (if you want to improve query parsing)
# nltk==3.8.1
# spacy==3.7.2

